requirements.txt
Flask==2.3.3
flask-cors==6.0.1
numpy==1.26.4
pandas==1.5.3
openpyxl==3.1.2
openai==1.100.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0